requirements.txt
wheel
transformers==4.39.3
accelerate==0.30.0
diffusers>=0.30.1
numpy==1.24.4
einops
ftfy
ipython
contexttimer
opencv-python-headless==4.10.0.84
imageio==2.33.1
imageio-ffmpeg==0.5.1
packaging
pandas
plotly
pre-commit
python-magic
scikit-image
sentencepiece
spacy==3.7.5
streamlit
timm==0.6.12
tqdm
torchmetrics
tiktoken
jsonlines
tensorboardX
